Modify Install.php to unset UUID for non-existing records
This commit introduces modifications to Install.php, in the Cli Installation Controller. It updates the logic to unset the UUID for non-existing records. At the same time, it simplifies a code block by directly calling the 'put' method without assigning its result to a variable. These changes aim to enhance data integrity and code expressiveness.