Refine code to dynamically retrieve filenames and paths
Made changes to the 'Install.php' in the CLI 'Install' controller to retrieve filenames and paths dynamically from the `Package.json` configuration. Now the files make use of the `System.Route.json` file to get updated paths. This new approach not only makes file handling in our packages more efficient but also boosts the overall maintainability of the system.