The provided commit message does not match the provided diffs. The appropriate message should be:
Commit message:
Add check for role system in Config

Added a check for the existence of a role system in Config.php. If a role system doesn't exist, it will create a basic one. This enhances the resilience of the system as it can create the required role systems if they're missing.