Add debug backtrace when log name not provided
Debug backtrace has been included in App.php to assist in identifying the source of a problem if the log name is not provided. This additional detail helps developers to solve any issues more quickly and efficiently by pinpointing the exact location in the code where the error has occurred.