Change installation command in Config module
The command used in the Config module for system installation has been adjusted. Previously, it was based on the Core's binary, which adds complexity. However, it has now been simplified to just using 'php', which makes the installation process more direct and less reliant on the binary status of Core. This should enhance the process efficiency and simplicity.