Add binary processing command in Config and Bin modules
Two new lines of code have been added in the Config and Bin modules to include binary processing. This change is expected to result in a more streamlined installation process. Instead of depending on the Core's binary, a simple 'php' command is enough to facilitate the process, thus enhancing efficiency and simplifying operations.